Cleanup and Bugfixes

+++ sub_process.py
@@ -1,22 +1,26 @@
-import config, os, pickle, datetime
+import config
+import os
+import pickle
+import datetime
 from leaderboard import *
 from models import User, Donor, DonorPhoneLog, CommittedDonation, EmailTemplate, DonorEmailLog, BulkEmailList
 from werkzeug.security import generate_password_hash, check_password_hash
 import email_service
 
+from exception import EntityNotFoundError
 
 # User Authentication Method Using MySQL Database.
 # This method returns the login_status, user_account_type and user full name if the user is authenticated ,
 # else it returns a None object
 # Return function is views.user_login_process()
+
+
 def authenticate(username, password):
-    print 'Authentication Module'
     login_status = False
     user_account_type = None
     user_full_name = None
     query_result = User.query.filter_by(username=username).first()
     if query_result is not None:
-        print query_result
         query_passwd = query_result.passwd
         query_account_status = query_result.account_status
         passwd_check = check_password_hash(query_passwd, password)
@@ -70,7 +74,6 @@ def write_user_log(str):
 # This method returns the list of all existing donors in the donor_details table in the donorworkflow database
 def get_donor_list():
     query_result = Donor.query.all()
-    query_length = len(query_result)
     return query_result
 
 
@@ -136,6 +139,8 @@ def get_pending_user_list():
 # once the Administrator has approved the User Account
 def approve_user(username):
     user_details = User.query.filter_by(username=username).first()
+    if user_details is None:
+        raise EntityNotFoundError
     user_details.account_status = 'Active'
     db.session.commit()
 
@@ -162,8 +167,9 @@ def approve_user(username):
 # once the Administrator has Rejected the User Account
 def reject_user(username):
     user_details = User.query.filter_by(username=username).first()
-    db.session.delete(user_details)
-    db.session.commit()
+    if user_details:
+        db.session.delete(user_details)
+        db.session.commit()
 
 
 # This method add the records of a new donor to the donor_details table of the
@@ -177,7 +183,6 @@ def register_new_donor(title, name, org, email, donor_contact, contact_person,
                          contact_date=contact_date, anonymous_select=anonymous_select,
                          volunteer_name=None))
     db.session.commit()
-    return None
 
 
 # This method add the records of a new donor to the DONOR DETAILS pickle file.
@@ -227,7 +232,7 @@ def donor_details_byid(donor_id):
 #         for i in range(0, obj_len):
 #             if int(pickle_obj[i]['id']) == int(donor_id):
 #                 return pickle_obj[i]
-#     return None
+#
 
 # This method return the phone conversation log with a Donor based on hid donor_id from the donor_contact_logs
 # table in the donorworkflow database.
@@ -264,7 +269,6 @@ def add_donor_phone_log(donor_id, contact_person, contact_date, contact_time, de
                                  contact_person=contact_person, donor_id=donor_id, remarks=remarks,
                                  details_shared=details_shared))
     db.session.commit()
-    return None
 
 
 # This method updates the DONOR_CONTACT_LOG pickle file with the details of the latest
@@ -300,6 +304,8 @@ def add_donor_phone_log(donor_id, contact_person, contact_date, contact_time, de
 # in the donor_details table in the donorworkflow database
 def update_donor_contact(donor_id, phone, email, org):
     user_details = Donor.query.filter_by(id=donor_id).first()
+    if user_details is None:
+        raise EntityNotFoundError
     user_details.phone = phone
     user_details.email = email
     user_details.org = org
@@ -361,8 +367,8 @@ def get_new_donor_list():
 
 
 def get_volunteer_list():
-    volunteer_list = User.query.filter_by(user_type='Volunteer').filter_by(account_status='Active').all()
-    print volunteer_list
+    volunteer_list = User.query.filter_by(
+        user_type='Volunteer').filter_by(account_status='Active').all()
     # user_account_file = config.USER_DETAILS_PICKLE_FILE
     # volunteer_list = []
     # if os.path.exists(user_account_file):
@@ -379,7 +385,8 @@ def get_volunteer_list():
 def allot_volunteer(donor_id, volunteer_name):
     donor = Donor.query.get(donor_id)
     donor.volunteer_name = volunteer_name
-    donor.donor_status = 'Allotted'
+    if donor.donor_status == 'New':  # if a donor is already allotted or committed
+        donor.donor_status = 'Allotted'
     db.session.commit()
     # donor_file = config.DONOR_DETAILS_PICKLE_FILE
     # donor_list = []
@@ -456,15 +463,12 @@ def alotted_donors_byid(username):
 
 def commit_donation(donor_id, commit_date, commit_time, commit_amt, currency,
                     payment_mode, remarks):
-    print 'Currency:' + currency
     db.session.add(CommittedDonation(donor_id=donor_id, commit_date=commit_date, commit_time=commit_time,
                                      commit_amt=commit_amt, currency=currency, payment_mode=payment_mode,
                                      remarks=remarks))
     donor = Donor.query.get(donor_id)
     donor.donor_status = 'Committed'
     db.session.commit()
-
-    return None
 
 
 def add_user_details_db(name, email, user_contact, username, password, user_type):
@@ -474,7 +478,6 @@ def add_user_details_db(name, email, user_contact, username, password, user_type
     db.session.add(User(username=username, full_name=name, email=email, phone=user_contact, user_type=user_type,
                         register_date=register_date, account_status=account_status, passwd=hashed_passwd))
     db.session.commit()
-    return None
 
 
 def find_donor(donor_id):
@@ -494,7 +497,6 @@ def add_new_emailtemplate(template_name, salutation, main_body, closing,
     db.session.add(EmailTemplate(template_name=template_name, salutation=salutation, main_body=main_body,
                                  closing=closing, signature_block=signature_block))
     db.session.commit()
-    return None
 
 
 def get_email_template_list():
@@ -504,14 +506,16 @@ def get_email_template_list():
 
 def get_email_template_obj(template_name, donor_obj, sender_name):
     if template_name != 'Manual Composition':
-        template_obj = EmailTemplate.query.filter_by(template_name=template_name).all()
+        template_obj = EmailTemplate.query.filter_by(
+            template_name=template_name).all()
         template_package = template_obj[0]
         if donor_obj.title is not None:
             donor_full_name = donor_obj.title + ' ' + donor_obj.name
         else:
             donor_full_name = donor_obj.name
         template_package.salutation = template_package.salutation + ' ' + donor_full_name
-        template_package.signature_block = sender_name + '\n' + template_package.signature_block
+        template_package.signature_block = sender_name + \
+            '\n' + template_package.signature_block
     else:
         template_package = EmailTemplate()
         template_package.template_name = 'Manual Composition'
@@ -525,9 +529,11 @@ def get_email_template_obj(template_name, donor_obj, sender_name):
 
 def get_bulk_email_template_obj(template_name, sender_name):
     if template_name != 'Manual Composition':
-        template_obj = EmailTemplate.query.filter_by(template_name=template_name).all()
+        template_obj = EmailTemplate.query.filter_by(
+            template_name=template_name).all()
         template_package = template_obj[0]
-        template_package.signature_block = sender_name + '\n' + template_package.signature_block
+        template_package.signature_block = sender_name + \
+            '\n' + template_package.signature_block
     else:
         template_package = EmailTemplate()
         template_package.template_name = 'Manual Composition'
@@ -544,8 +550,6 @@ def transmit_indl_email(donor_id, contact_person, contact_date, contact_time, sa
     donor = find_donor(donor_id)
     contact_date = datetime.datetime.strptime(contact_date, '%m-%d-%Y')
     contact_date = contact_date.date().strftime("%Y-%m-%d")
-    print(donor.email, contact_person, contact_date, contact_time, salutation,
-          main_body, closing, signature)
     message_obj, mail_code, msg_body = email_service.outgoing_mail_process(donor_email=donor.email,
                                                                            contact_person=contact_person,
                                                                            contact_date=contact_date,
@@ -556,23 +560,21 @@ def transmit_indl_email(donor_id, contact_person, contact_date, contact_time, sa
     db.session.add(DonorEmailLog(contact_date=contact_date, contact_time=contact_time, contact_person=contact_person,
                                  donor_id=donor_id, main_body=main_body, full_email=msg_body, mail_code=mail_code))
     db.session.commit()
-    return None
 
 
-def add_donor_bulk_email_list(donor_id):
-    current_username = session['logged_username']
-    existing_list_check = BulkEmailList.query.filter_by(donor_id=donor_id).filter_by(username=current_username).first()
-    print(existing_list_check)
-
+def add_donor_bulk_email_list(donor_id, sender_username):
+    existing_list_check = BulkEmailList.query.filter_by(
+        donor_id=donor_id).filter_by(username=sender_username).first()
 
     if existing_list_check is None:
-        db.session.add(BulkEmailList(donor_id=donor_id,username=current_username))
+        db.session.add(BulkEmailList(
+            donor_id=donor_id, username=sender_username))
         db.session.commit()
-    return None
 
 
 def get_bulk_email_donor_details(sender_username):
-    donor_list_query = BulkEmailList.query.filter_by(username=sender_username).all()
+    donor_list_query = BulkEmailList.query.filter_by(
+        username=sender_username).all()
     donor_list = []
     for donors in donor_list_query:
         donor_list.append(int(donors.donor_id))
@@ -590,8 +592,6 @@ def transmit_bulk_email(bulk_email_donor_details, contact_person, contact_date, 
         else:
             donor_full_name = donors.name
         temp_salutation = salutation + ' ' + donor_full_name
-        print(donors.email, contact_person, contact_date, contact_time, salutation,
-              main_body, closing, signature)
         message_obj, mail_code, msg_body = email_service.outgoing_mail_process(donor_email=donors.email,
                                                                                contact_person=contact_person,
                                                                                contact_date=contact_date,
@@ -600,7 +600,6 @@ def transmit_bulk_email(bulk_email_donor_details, contact_person, contact_date, 
                                                                                main_body=main_body,
                                                                                closing=closing, signature=signature)
         db.session.add(DonorEmailLog(contact_date=contact_date, contact_time=contact_time, contact_person=contact_person,
-                          donor_id=donors.id, main_body=main_body, full_email=msg_body, mail_code=mail_code))
+                                     donor_id=donors.id, main_body=main_body, full_email=msg_body, mail_code=mail_code))
         db.session.commit()
         BulkEmailList.query.filter_by(username=sender_username).delete()
-    return None
